clean code

+++ app/src/main/java/com/trak/mem/scene/play/component/HeaderView.kt
@@ -11,16 +11,24 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.trak.mem.R
 import com.trak.mem.common.component.TextIconView
 
+/**
+ * Header View
+ *
+ * @param clickLimit - click limit when null no limit
+ * @param timeLimit - time limit when null no limit
+ * @param onReset - on reset action
+ * @param tint - tint color
+ * @param modifier -modifier
+ */
 @Composable
 fun HeaderView(
     clickLimit: Int?,
-    timeLimit: Int?,
+    timeLimit: Long?,
     onReset: () -> Unit,
     tint: Color,
     modifier: Modifier = Modifier
